Refactor updater to use library
